--- src/main/java/ru/trofimov/app/exceptions/UserAlreadyContainsException.java
@@ -1,7 +1,0 @@
-package ru.trofimov.app.exceptions;
-
-public class UserAlreadyContainsException extends RuntimeException {
-    public UserAlreadyContainsException(String message) {
-        super(message);
-    }
-}
