--- src/main/java/ru/trofimov/app/exceptions/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.trofimov.app.exceptions;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
