+++ src/main/java/ru/trofimov/app/operations/OperationFactory.java
@@ -18,4 +18,4 @@ public class OperationFactory {
     public Operation getOperation(OperationTypes operationType) {
         return context.getBean(operationType.getOperationClass());
     }
- }
+}
