+++ src/main/java/ru/trofimov/app/operations/impl/GetAccountOperation.java
@@ -6,7 +6,6 @@ import ru.trofimov.app.entity.Account;
 import ru.trofimov.app.operations.Operation;
 import ru.trofimov.app.service.AccountService;
 
-import javax.security.auth.login.AccountNotFoundException;
 import java.util.Scanner;
 
 @Component
@@ -25,11 +24,7 @@ public class GetAccountOperation implements Operation {
         int accountId = scanner.nextInt();
         scanner.nextLine();
 
-        try {
-            Account account = accountService.getAccount(accountId);
-            System.out.println("Account information: " + account);
-        } catch (AccountNotFoundException e) {
-            System.out.println(e.getMessage());
-        }
+        Account account = accountService.getAccount(accountId);
+        System.out.println("Account information: " + account);
     }
 }
